@@ -1,6 +1,6 @@
 {
   "name": "ngx-customscrollbar",
-  "version": "2.0.1",
+  "version": "2.0.2",
   "scripts": {
     "ng": "ng",
     "demo": "ng build ngx-customscrollbars && npm start",
@@ -14,45 +14,45 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "8.1.1",
-    "@angular/common": "8.1.1",
-    "@angular/compiler": "8.1.1",
-    "@angular/core": "8.1.1",
-    "@angular/forms": "8.1.1",
-    "@angular/material": "8.0.2",
-    "@angular/platform-browser": "8.1.1",
-    "@angular/platform-browser-dynamic": "8.1.1",
-    "@angular/router": "8.1.1",
-    "rxjs": "6.5.2",
-    "tslib": "1.10.0",
-    "zone.js": "0.9.1"
+    "@angular/animations": "8.2.14",
+    "@angular/common": "8.2.14",
+    "@angular/compiler": "8.2.14",
+    "@angular/core": "8.2.14",
+    "@angular/forms": "8.2.14",
+    "@angular/material": "8.2.3",
+    "@angular/platform-browser": "8.2.14",
+    "@angular/platform-browser-dynamic": "8.2.14",
+    "@angular/router": "8.2.14",
+    "rxjs": "6.5.3",
+    "tslib": "1.10.0"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "0.801.1",
-    "@angular-devkit/build-ng-packagr": "0.801.1",
-    "@angular/cdk": "8.0.2",
-    "@angular/cli": "8.1.1",
-    "@angular/compiler-cli": "8.1.1",
-    "@angular/language-service": "8.1.1",
-    "@types/jasmine": "3.3.13",
-    "@types/jasminewd2": "2.0.6",
-    "@types/node": "12.6.2",
-    "codelyzer": "5.1.0",
-    "core-js": "3.1.4",
+    "@angular-devkit/build-angular": "0.803.20",
+    "@angular-devkit/build-ng-packagr": "0.803.20",
+    "@angular/cdk": "8.2.3",
+    "@angular/cli": "8.3.20",
+    "@angular/compiler-cli": "8.2.14",
+    "@angular/language-service": "8.2.14",
+    "@types/jasmine": "3.5.0",
+    "@types/jasminewd2": "2.0.8",
+    "@types/node": "12.12.14",
+    "codelyzer": "5.2.0",
+    "core-js": "3.4.5",
     "cpx": "1.5.0",
     "jasmine-core": "3.4.0",
     "jasmine-spec-reporter": "4.2.1",
-    "karma": "4.2.0",
-    "karma-chrome-launcher": "3.0.0",
-    "karma-coverage-istanbul-reporter": "2.0.5",
+    "karma": "4.4.1",
+    "karma-chrome-launcher": "3.1.0",
+    "karma-coverage-istanbul-reporter": "2.1.1",
     "karma-jasmine": "2.0.1",
     "karma-jasmine-html-reporter": "1.4.2",
-    "ng-packagr": "5.3.0",
+    "ng-packagr": "5.7.1",
     "protractor": "6.0.0",
-    "ts-node": "8.3.0",
-    "tsickle": "0.35.0",
+    "ts-node": "8.5.2",
+    "tsickle": "0.37.1",
     "tslib": "1.10.0",
-    "tslint": "5.18.0",
-    "typescript": "3.4.5"
+    "tslint": "5.20.1",
+    "typescript": "^3.5.3",
+    "zone.js": "^0.9.1"
   }
 }
